==> code/01_cohorting.qmd <==
---
title: "01_cohort_identification"
author: Nick Ingraham
date: today
execute: 
  echo: false
format: 
  html:
    embed-resources: true
    number-sections: true
    toc: true
    html-q-tags: true
    code-fold: true
editor: source
editor_options: 
  chunk_output_type: console
---


# Cohort identification

```{r}

start.time.full <- Sys.time()

#~~~~~~~~~~
# ID cohort
#~~~~~~~~~~

# Main Respiratory Support Table
#~~ using readr or fst or arrow (parquet) package depending on the file type. (THESE WERE SET DURING SET UP ^^ )
clif_respiratory_support_start <- 
  read_data("clif_respiratory_support") |> 
  dplyr::select(hospitalization_id, device_category, tracheostomy, recorded_dttm) |> 
# Just getting those ventilated or trach
  filter(str_to_lower(device_category) == "imv" | tracheostomy == 1) |> 
  mutate(across(where(is.character), str_to_lower)) 

    
clif_hospitalization_start <- 
  read_data("clif_hospitalization") |> 
  dplyr::select(hospitalization_id, admission_dttm, discharge_dttm, age_at_admission, patient_id, 
         # admission_type_name, 
         # admission_type_category, 
         discharge_name, discharge_category) |> 
  mutate(admission_dttm = as.POSIXct(admission_dttm, tz = "UTC")) |> 
  mutate(admission_dttm = ymd_hms(admission_dttm)) |> 
  mutate(discharge_dttm = as.POSIXct(discharge_dttm, tz = "UTC")) |>       
  mutate(discharge_dttm = ymd_hms(discharge_dttm)) |>       
  # unfactor things to ensure everything is lowercase
  mutate(across(where(is.factor), as.character)) |> 
  # get everything to lower
  mutate(across(where(is_character), str_to_lower)) 


# getting a keep_list of hospitalization_ids that have vent or trach during their encounter... we will use this right_join the dfs below
# Update 10/2024... 
#     need to just use patient_id for now 
#     because we need to get hsopitalizations_joined_id first 
#     before we pair down to the hospitalizations we want to keep 
df_cohort_keep_start <- clif_respiratory_support_start |> 
  

# Exclusion based on date
  left_join(clif_hospitalization_start) |>
  filter(
    lubridate::date(admission_dttm) >= start_date,
    lubridate::date(admission_dttm) <= end_date,
# age exclusion
    age_at_admission >= 18,
# missing data exclusion
    !is.na(age_at_admission)
    ) |>  
  dplyr::select(patient_id) |> 
  distinct()


# shrinking down to just the patients that meet criteria
clif_hospitalization_start <- clif_hospitalization_start |> 
  right_join(df_cohort_keep_start) 


```



## Fixing Hospitalization

```{r}
#Create an Hospital Block ID - This is to Identify Continuous Hospitalizations When Patients Are Transferred Between Hospitals in One Health System
#This code is intended be robust to various ways encounters may be coded in CLIF databases
hospital_blocks <- clif_hospitalization_start |> 
  dplyr::select(patient_id, hospitalization_id, admission_dttm, discharge_dttm) |>
  arrange(patient_id, admission_dttm) |>
  collect()

#Identify Admissions That Occur Within 3 Hours of a Discharge (Will Consider Those Linked and as Part of One Continuous Encounter)
#Use Data Table for Speed
linked_encounters <- setDT(hospital_blocks)
#Create a Variable for the time of the next admission and time of previous discharge
linked_encounters[, ':=' (next_admit_dttm = data.table::shift(admission_dttm, n=1, type = "lead")), by = patient_id]
linked_encounters[, ':=' (prev_dc_dttm = data.table::shift(discharge_dttm, n=1, type = "lag")), by = patient_id]
#Calculates Time Between Discharge and Next Admit
linked_encounters[, next_diff_time := difftime(next_admit_dttm, discharge_dttm, units = "hours")]
linked_encounters[, prev_diff_time := difftime(admission_dttm, prev_dc_dttm, units = "hours")]

#Now Create Variable Indicating a Linked Encounter (next_admit-dc time <6 hours or prev_dc-admint <6 hours)
linked_encounters[, linked := fcase(
  (next_diff_time <6 | prev_diff_time <6), 1)]
#Filter to Only Linked Encounters and number them
linked_encounters <- linked_encounters[linked==1]
#This Identifies the First Encounter in a Series of Linked Encounters
linked_encounters[, first_link := fcase(
  (rowid(linked)==1 | (next_diff_time<6 & prev_diff_time>6)), 1
), by = patient_id]

#Now Numbers Encounters, easier in dplyr
#Filter to Just First Links, Number them and then Remerge with linked encounters
temp <- as_tibble(linked_encounters) |>
  filter(first_link==1) |>
  group_by(patient_id) |>
  mutate(link_group=row_number()) |>
  ungroup() |>
  dplyr::select(hospitalization_id, link_group) 
linked_encounters <- as_tibble(left_join(linked_encounters, temp)) |>
  fill(link_group, .direction = c("down")) |>
  #Create a Variable Indicating Which Number of LIinked Encounter the Encounter is
  group_by(patient_id, link_group) |>
  mutate(link_number=row_number()) |>
  ungroup() |>
  dplyr::select(hospitalization_id, linked, link_number)
rm(temp)

#Now Join Back to Hospitalization Table
clif_hospitalization_start <- clif_hospitalization_start |>
  left_join(linked_encounters) |>
  mutate(linked=if_else(is.na(linked), 0, linked)) |>
  compute()

#Pull Out the Any Linked Encounter that Is NOt the First Encounter and Assign Each Encounter an Encounter Block ID in the Original clif_hospitalization table
df_link <- clif_hospitalization_start |>
  filter(link_number>1) |>
  collect()

clif_hospitalization <- clif_hospitalization_start |>
  group_by(patient_id) |>
  arrange(patient_id, admission_dttm) |>
  #Remove Link Numbers that Are Not First in Link Encounter
  filter(link_number==1 | is.na(link_number)) |>
  #Make Encounter Blocks
  collect() |>
  mutate(encounter_block=row_number()) |>
  rowbind(df_link, fill = TRUE) |> #Bring Back in Link Numbers >1
  group_by(patient_id) |> arrange(patient_id, admission_dttm) |>
  fill(encounter_block, .direction = "down") |>
  ungroup()|>
  #Finally, for Linked Encounters Identify 'Final_admit_date' and 'final_dc_date' which are the first and last dates of a link block
  group_by(patient_id, encounter_block) |>
  mutate(final_admission_dttm=fcase(
    row_number()==1, admission_dttm
  )) |>
  mutate(final_discharge_dttm=fcase(
    row_number()==n(), discharge_dttm
  )) |>
  # mutate(final_admission_type_name=fcase(
  #   row_number()==1, admission_type_name
  # )) |>
  mutate(final_discharge_name=fcase(
    row_number()==n(), discharge_name
  )) |>
  # mutate(final_admission_type_category=fcase(
  #   row_number()==1, admission_type_category
  # )) |>
  mutate(final_discharge_category=fcase(
    row_number()==n(), discharge_category
  )) |>
  mutate(clif_hospitalizations_joined_id = cur_group_id()) |> 
  fill(starts_with("final_"), .direction = 'updown') |>
  
  dplyr::select(
    clif_hospitalizations_joined_id,
    hospitalization_id,
    patient_id,
    encounter_block,
    age_at_admission,
    admission_dttm = final_admission_dttm,
    discharge_dttm = final_discharge_dttm,
    # admission_type_name = final_admission_type_name,
    discharge_name = final_discharge_name,
    # admission_type_category = final_admission_type_category,
    discharge_category = final_discharge_category
  ) |> 
  ungroup() |> 
  as_arrow_table() |> 
  collect()



```

### Final cohorting with fixed clif_hospitalizations_joined_id
```{r}

clif_patient_start <-  read_data("clif_patient") |> 
  dplyr::select(patient_id, sex_category) |> 
  filter(str_detect(str_to_lower(sex_category), "male|female")) |> 
  dplyr::select(patient_id) |> 
  distinct()


# need to get the joined encounters taht we need to keep... then we will make a final keep with 
#     clif_hospitalizations_joined_id, hospitalization_id, patient_id
df_cohort_keep_encounters <- 
  # These are just patient_ids 
  df_cohort_keep_start |> 

# redo inclusion/exclusion with new admit dates and d/c dates
# Getting hospitaliaztion_ids and such with new variables made above
# Exclusion based on date
  left_join(clif_hospitalization) |>
  inner_join(clif_patient_start) |> 
  filter(
    lubridate::date(admission_dttm) >= start_date,
    lubridate::date(admission_dttm) <= end_date,
# age exclusion
    age_at_admission >= 18,
# missing data exclusion
    !is.na(age_at_admission)
    ) |>  
# these are JUST hospitalizations with imv or trach... we are using this to ID the joined_ids that we need to keep
  inner_join(clif_respiratory_support_start |> dplyr::select(hospitalization_id) |> distinct()) |> 
  dplyr::select(clif_hospitalizations_joined_id) |> 
  distinct()

df_cohort_keep <- clif_hospitalization |> 
  inner_join(df_cohort_keep_encounters) |> 
  mutate(patient_id = as.character(patient_id)) |> 
  mutate(hospitalization_id = as.character(hospitalization_id)) |> 
  dplyr::select(clif_hospitalizations_joined_id, hospitalization_id, patient_id) |> 
  distinct()
  
# dropping hospitals with < 100 encounters
hospital_keep <- read_data("clif_adt") %>%
  mutate(across(where(is.character), str_to_lower)) |> 

  right_join(df_cohort_keep) |> 
  mutate(location_category = factor(location_category,
                                    levels = (c("icu", "stepdown", "ward", "ed", "procedural", "other")), ordered = TRUE)) |> 
  filter(location_category == "icu") |> 
  group_by(clif_hospitalizations_joined_id) |> 
  arrange(clif_hospitalizations_joined_id, location_category, in_dttm) |> 
  
  # first one will be ICU
  slice_head(n=1) |> 
  ungroup() |> 
  dplyr::select(clif_hospitalizations_joined_id, hospital_id) |> 
  mutate(hospital_id_count = n(),
         .by = "hospital_id") |> 
  filter(hospital_id_count > 100)

# Count the number of encounters per hospital_id and assign "Hospital 1", "Hospital 2", etc.
hospital_order <- hospital_keep |> 
  count(hospital_id, name = "encounters") %>%
  arrange(desc(encounters)) %>%
  mutate(hospital_label = paste0("Hospital ", site_name," ", row_number()))

# use this in analytic sets after we ID the ICU/VENT hosp to filter out hospitals that don't have enough traffic
# filter(hospital_id %in% hospital_filter)
hospital_filter <- hospital_order |> pull()

df_cohort_keep_encounters <- df_cohort_keep_encounters |> 
  # drop those at hospitals with < 100
  inner_join(hospital_keep |> select(clif_hospitalizations_joined_id))
  
df_cohort_keep <- df_cohort_keep |> 
  inner_join(df_cohort_keep_encounters)

# narrowing down clif_hospitaliation to the appropriate cohort
clif_hospitalization <- clif_hospitalization |> 
  inner_join(df_cohort_keep_encounters)
  
  
  
```


### Importing function

```{r}
#~~~~~~~~~~
# Importing function
#~~~~~~~~~~

# Function for importing files easily just with the file of the clif_table
import_df <- function(x) {
  # Construct the file path
  file_path <- paste0(tables_path,"/", x, ".", file_type)
  
  # Dynamically call the appropriate read function based on file_type
  
  if (file_type == "parquet") {
    
    read_function <- get(paste0("open_dataset")) 
    
    df <- read_function(file_path) |>
      
      mutate(across(ends_with("_id"), as.character)) |> 
      
      # helps with glitch in open_dataset
      mutate(across(where(is.character), as.character)) |>
      
      # # helps with glitch in open_dataset
      # mutate(across(where(is.double), as.integer)) |> 
      
      # right join early so files are smaller
      right_join(df_cohort_keep) |> 
      
      # unfactor things to ensure everything is lowercase
      mutate(across(where(is.factor), as.character)) |> 
      
      # get everything to lower
      mutate(across(where(is_character), str_to_lower)) |> 
      
      # only need if using open dataset
      collect()
    
  } else {
    
    read_function <- get(paste0("read_", file_type)) 
    
    df <- read_function(file_path) |>
      mutate(across(ends_with("_id"), as.character)) |> 
      
      # right join early so files are smaller
      right_join(df_cohort_keep) |> 
      
      # unfactor things to ensure everything is lowercase
      mutate(across(where(is.factor), as.character)) |> 
      
      # get everything to lower
      mutate(across(where(is_character), str_to_lower)) 
    
  }
  
  
  # Find all datetime variables containing 'dttm'
  datetime_vars <- names(df)[grepl("dttm", names(df))]
  
  # Check and convert datetime variables
  for (var in datetime_vars) {
    if (!inherits(df[[var]], "POSIXct")) {
      
      df[[var]] <- as.POSIXct(df[[var]], tz = "UTC")
      
      # Attempt to parse datetime using a common format, adjust based on your actual format
      df[[var]] <- ymd_hms(df[[var]], quiet = TRUE)
      
      # Check if it just turned things into NA
      if (is.na(ffirst(df[[var]], na.rm = TRUE))) {
        stop(paste("Conversion failed for:", var,"\n","\nPlease look at your clif data for   ", var, "\nmake sure its POSIXct format!!"))
        
        
        # Check if conversion failed (if it's still not POSIXct)
        if (!inherits(df[[var]], "POSIXct")) {
          stop(paste("Conversion failed for:", var,"\n","\nPlease look at your clif data for   ", var, "\nmake sure its POSIXct format!!"))
          
        }
        
      }
    }
    print(paste("Woohoo",var,"is in the correct dttm format!!"))
  }
  
  # Join with cohort data and glimpse the result
  df %>% glimpse()
  
  # Assign the dataframe to a new variable in the global environment named after the file_name
  assign(x, df, envir = .GlobalEnv)
}




# List all objects in the global environment
all_objects <- ls()

# Identify objects ending with "_start"
df_start_objects <- grep("_start$", all_objects, value = TRUE)

# # Remove identified objects
# rm(list = df_start_objects)
# gc()

```



### Importing tables

```{r}

#~~~~~~~~~~
# Import data and cohort tables
#~~~~~~~~~~


# Resp support table cohort
clif_respiratory_support <- import_df("clif_respiratory_support") |> 
      # Check if Variables match!!
      # You will get an error if you do not have all the relevant variables needed
      dplyr::select(
        clif_hospitalizations_joined_id, hospitalization_id,
        recorded_dttm,
        device_name,
        device_category,
        mode_name,
        mode_category,
        fio2_set,
        lpm_set,
        tidal_volume_set,
        resp_rate_set,
        pressure_control_set, # not used
        pressure_support_set,
        # flow_rate_set,  # not used
        # peak_inspiratory_pressure_set, # not used and only in infants
        inspiratory_time_set, # not used
        peep_set,
        starts_with("tidal_volume_obs"),
        plateau_pressure_obs,
        peak_inspiratory_pressure_obs,
        starts_with("mean_airway_pressure"),  # in case ppl don't have this yet.. prevents from breaking
        # peep_obs, # not used and not in adult set 
        starts_with("minute_vent_obs"),
        tracheostomy) |> 
  mutate(across(ends_with("_set"), as.double)) |> 
  mutate(across(ends_with("_obs"), as.double))


# ADT File to merge with patient so we can get hospital ID
clif_adt <- import_df("clif_adt") |> 
      # Check if Variables match!!
      dplyr::select(
        clif_hospitalizations_joined_id, hospitalization_id,
        location_name,
        location_category,
        in_dttm,
        out_dttm,
        hospital_id)|> 
  filter(!is.na(location_category))


# Count the number of encounters per hospital_id and assign "Hospital 1", "Hospital 2", etc.
hospital_order <- clif_adt %>%
  count(hospital_id, name = "encounters") %>%
  arrange(desc(encounters)) %>%
  mutate(hospital_label = paste0("Hospital ", site_name," ", row_number()))

clif_adt <- clif_adt |> 
  left_join(hospital_order %>% dplyr::select(hospital_id, hospital_label), by = "hospital_id") |> 
  dplyr::select(-hospital_id) |> 
  rename(hospital_id = hospital_label) 
  
  

# Vitals to merge with patient so we can get BMI, height, spo2 information
clif_vitals <- import_df("clif_vitals") |> 
      # Check if Variables match!!
      dplyr::select(
        clif_hospitalizations_joined_id, hospitalization_id,
        recorded_dttm,
        vital_name,
        vital_category,
        vital_value
        # meas_site_name
        ) |> 
  filter(!is.na(vital_category))


# GCS for lap2
clif_gcs <- import_df("clif_patient_assessments") |> 
  filter(assessment_category == "gcs_total") |> 
      # Check if Variables match!!
      dplyr::select(
        clif_hospitalizations_joined_id, hospitalization_id,
        recorded_dttm,
        gcs_total = numerical_value
        )

# Labs to merge with patient so we can get pao2 all lap2 needs
clif_labs <- import_df("clif_labs") |> 
      # Check if Variables match!!
      dplyr::select(
        clif_hospitalizations_joined_id, hospitalization_id,
        lab_order_dttm,
        lab_result_dttm,
        lab_collect_dttm,
        lab_name,
        lab_category,
        # lab_order_category,
        lab_value_numeric,
        reference_unit) |> 
  arrange(clif_hospitalizations_joined_id, hospitalization_id, lab_collect_dttm, lab_category, lab_value_numeric) |> 
  
  # get rid of duplicates
  #     will take the first non-missing lab_value_numeric
  distinct(clif_hospitalizations_joined_id, hospitalization_id, lab_collect_dttm, lab_category, .keep_all = TRUE) |> 
    # get rid of those with NO tidal volumes ever
  group_by(clif_hospitalizations_joined_id) |> 
  filter(!is.na(lab_category)) |> 
  ungroup() 
  

# Demographics 
clif_patient <- import_df("clif_patient") |> 
      # Check if Variables match!!
      dplyr::select(
        clif_hospitalizations_joined_id, hospitalization_id,
        patient_id,
        sex_category,
        race_category,
        ethnicity_category,
        death_dttm) |> 
  arrange(patient_id, sex_category) |> 
  # account for possible "male" and "unknown" documentation. very race.  taking male > unknown 
  # distinct will slice the top one.
  distinct(clif_hospitalizations_joined_id, hospitalization_id, patient_id, .keep_all = TRUE) |> 
  filter(sex_category != "unknown")


# THIS IS DONE ALREADY ABOVE!!
# # Encounter Information 
# clif_hospitalization <- import_df("clif_hospitalization") |>
#       # Check if Variables match!!
#       dplyr::select(clif_hospitalizations_joined_id, hospitalization_id,
#              patient_id,
#              # admission_type_category, # don't have this yet in CLIF
#              admission_dttm,
#              discharge_dttm,
#              discharge_category,
#              discharge_name,
#              age_at_admission)



# combining them to one df
clif_demographics_combined <- clif_patient |> 
  left_join(clif_hospitalization) |> dplyr::select(-hospitalization_id) |> distinct()



# # Procedures (don't need for now, just for checking trach data at some point)
# clif_procedures <- import_df("clif_procedures") |>
#       # Check if Variables match!!
#       c(clif_hospitalizations_joined_id, hospitalization_id,
#         procedure_name,
#         start_time,
#         hcup_category)     # UMN variable... may want to implement in CLIF... just need to link HCUP table to PX codes



```

# Import QA checks

```{r}

# Check fio2_set
fio2_mean <- mean(clif_respiratory_support$fio2_set, na.rm = TRUE) 

# fixing if its less than one
# You will get a warning but it will be fixed on its own with IF statement
if(fio2_mean < 1){ 
  fio2warning()
  clif_respiratory_support <- clif_respiratory_support |> mutate(fio2_set = fio2_set * 100)
}


# Check Values of labs and vitals
if(is.character(clif_labs$lab_value_numeric)){
  labwarning()
  clif_labs <- clif_labs |> 
    mutate(lab_value_numeric = as.numeric(parse_number(lab_value_numeric)))
}


# Check Values of labs and vitals
if(is.character(clif_vitals$vital_value)){
  vitalwarning()
  clif_vitals <- clif_vitals |> 
    mutate(vital_value = as.numeric(parse_number(vital_value)))
}

```



# Cleaning Data

## Hourly Sequence

```{r}
#| label: cleaning up vent data


#~~~~~~~~~~~~~~~~
##~~ getting an hour sequence so we can fill in the gaps
#~~~~~~~~~~~~~~~~
## This is just encounter ID and recorded times at xx:59:59
## data that occurs last in the hour when there are multiple data points 
## in the end ... we will want this to be the data we use to fill the next hour... if time is NOT unified... you could have hour sequence that is 12:01, 1:01 everywhere and even when there IS data you risk filling in from the hour before and not getting the NEW data during that hour.
# if we set all the new seq hours to 59:59 then you can fill those in without risking other data when you fill in and do distinct (take the first of the hour for everything).  Remember.  Even hours with 1 data will have a new hour seq row that may be before or after the data... so doing the 59:59 puts it at the end!!!
# 




hour_sequence <- clif_respiratory_support |> 
  group_by(clif_hospitalizations_joined_id)  |> 

  reframe(recorded_dttm = seq(fmin(recorded_dttm), fmax(recorded_dttm), by = "1 hour")) |>
  # Adjust to the last second of the hour using lubridate's floor_date
  mutate(recorded_dttm = floor_date(recorded_dttm, "hour") + minutes(59) + seconds(59)) |> 
  # Create date and hour columns efficiently
  mutate(recorded_date = as_date(recorded_dttm),
         recorded_hour = hour(recorded_dttm)) |> 
  ungroup()

```

## Quality Check & Clean + Waterfall

```{r}
#~~~~~~~~~~~~~~~~
##~~ Quick QA and fixing missing values throughout
#~~~~~~~~~~~~~~~~
ni_tic()
df_resp_support_1  <- clif_respiratory_support |> 
  dplyr::select(clif_hospitalizations_joined_id, hospitalization_id, recorded_dttm, device_category, device_name, mode_category, mode_name, 
         fio2_set, lpm_set, tidal_volume_set, peep_set,  pressure_support_set, resp_rate_set, tracheostomy, 
         starts_with("tidal_volume_obs"), peak_inspiratory_pressure_obs, 
         starts_with("minute_vent_obs"), 
         plateau_pressure_obs, starts_with("mean_airway_pressure_obs") 
  )

if ("mean_airway_pressure_obs" %in% names(df_resp_support_1)){
  df_resp_support_1 <- df_resp_support_1 |> 
    mutate(
          
    # mean_airway_pressure_obs
    mean_airway_pressure_obs = fcase(mean_airway_pressure_obs > 60, NA_real_, rep_len(TRUE, length(mean_airway_pressure_obs)), mean_airway_pressure_obs),
    mean_airway_pressure_obs = fcase(mean_airway_pressure_obs <  0, NA_real_, rep_len(TRUE, length(mean_airway_pressure_obs)), mean_airway_pressure_obs)
    
    )
  print("QC for MAIRP done!!")
}
  
  
df_resp_support_1 <- df_resp_support_1 |> 
  mutate(
    # fio2_set
    fio2_set = fcase(fio2_set > 100, NA_real_, rep_len(TRUE, length(fio2_set)), fio2_set),
    fio2_set = fcase(fio2_set <  21, NA_real_, rep_len(TRUE, length(fio2_set)), fio2_set),
    
    # Set tidal_volume_set
    tidal_volume_set = fcase(tidal_volume_set > 2500, NA_real_, rep_len(TRUE, length(tidal_volume_set)), tidal_volume_set),
    tidal_volume_set = fcase(tidal_volume_set <   50, NA_real_, rep_len(TRUE, length(tidal_volume_set)), tidal_volume_set),
    
    # peep_set
    peep_set = fcase(peep_set > 30, NA_real_, rep_len(TRUE, length(peep_set)), peep_set),
    peep_set = fcase(peep_set <  0, NA_real_, rep_len(TRUE, length(peep_set)), peep_set),
    
    # pressure_support_set (sometimes APRV may be in here ... so limit ~ 50??)
    pressure_support_set = fcase(pressure_support_set > 50, NA_real_, rep_len(TRUE, length(pressure_support_set)), pressure_support_set),
    pressure_support_set = fcase(pressure_support_set <  0, NA_real_, rep_len(TRUE, length(pressure_support_set)), pressure_support_set),
    
    # resp_rate_set
    resp_rate_set = fcase(resp_rate_set > 60, NA_real_, rep_len(TRUE, length(resp_rate_set)), resp_rate_set),
    resp_rate_set = fcase(resp_rate_set <  0, NA_real_, rep_len(TRUE, length(resp_rate_set)), resp_rate_set),
    
    # tidal_volume_obs
    # tidal_volume_obs = fcase(tidal_volume_obs > 2500, NA_real_, rep_len(TRUE, length(tidal_volume_obs)), tidal_volume_obs),
    # tidal_volume_obs = fcase(tidal_volume_obs <    0, NA_real_, rep_len(TRUE, length(tidal_volume_obs)), tidal_volume_obs),
    
    # peak_inspiratory_pressure_obs
    peak_inspiratory_pressure_obs = fcase(peak_inspiratory_pressure_obs > 60, NA_real_, rep_len(TRUE, length(peak_inspiratory_pressure_obs)), peak_inspiratory_pressure_obs),
    peak_inspiratory_pressure_obs = fcase(peak_inspiratory_pressure_obs <  0, NA_real_, rep_len(TRUE, length(peak_inspiratory_pressure_obs)), peak_inspiratory_pressure_obs),
    
    # # minute_vent_obs
    # minute_vent_obs = fcase(minute_vent_obs > 30, NA_real_, rep_len(TRUE, length(minute_vent_obs)), minute_vent_obs),
    # minute_vent_obs = fcase(minute_vent_obs <  0, NA_real_, rep_len(TRUE, length(minute_vent_obs)), minute_vent_obs),
    # 

    ) |>
  
  # getting data and hour information
  mutate(recorded_date = date(recorded_dttm),
         recorded_hour = hour(recorded_dttm)) |> 
  
  # getting hospital ID for each hour
  dplyr::left_join( #tidy_table doesn't like it when you use join_by() with between (Dropped tidy_table 2_2024) 
    clif_adt |> 
      dplyr::select(clif_hospitalizations_joined_id, hospital_id, location_name, location_category, in_dttm, out_dttm),
    by = join_by(clif_hospitalizations_joined_id, between(recorded_dttm, in_dttm, out_dttm))
  ) |> 

  # order for filling things in
  arrange(clif_hospitalizations_joined_id, recorded_dttm) |> 
  

  
  # Fixing when: the mode and category are there with device_name and device_cat not filled in.  fixing with the below
  mutate(
    device_category = 
      fcase(
        is.na(device_category) & is.na(device_name) &
          str_detect(mode_category, "assist control-volume control|simv|pressure control"),
        "imv",
            rep_len(TRUE, length(device_category)), device_category
      ),
    device_name = 
      fcase(
        str_detect(device_category, "imv") & is.na(device_name) &
          str_detect(mode_category, "assist control-volume control|simv|pressure control"),
        "mechanical ventilator",
        rep_len(TRUE, length(device_name)), device_name
        
      ),
  ) |>
  
  # fixing other vent things
  #     If device before is VENT + normal vent things ... its VENT too 
  mutate(device_category = fcase(is.na(device_category) & 
                                     lag(device_category == "imv") & 
                                     tidal_volume_set > 1 & 
                                     resp_rate_set > 1 & 
                                     peep_set > 1, 
                                   "imv", 
                                   rep_len(TRUE, length(device_category)), device_category)) |>
  
  #     If device after is VENT + normal vent things ... its VENT too 
  mutate(device_category = fcase(is.na(device_category) & 
                                     lead(device_category == "imv") & 
                                     tidal_volume_set > 1 & 
                                     resp_rate_set > 1 & 
                                     peep_set > 1, 
                                   "imv", 
                                   rep_len(TRUE, length(device_category)), device_category)) |>
  
  # same as above for device_name ^^^^^^^^^^^
  mutate(device_name = fcase(is.na(device_name) & lag(device_category == "imv") & tidal_volume_set > 1 & resp_rate_set > 1 & peep_set > 1, 
                             "mechanical ventilation", 
                             rep_len(TRUE, length(device_name)), device_name)) |> 
  
  mutate(device_name = fcase(is.na(device_name) & lead(device_category == "imv") & tidal_volume_set > 1 & resp_rate_set > 1 & peep_set > 1, 
                             "mechanical ventilation", 
                             rep_len(TRUE, length(device_name)), device_name)) |> 
  
  
  # doing this for BiPAP as well 
  mutate(device_category = fcase(is.na(device_category) & 
                                     lag(device_category == "nippv") & 
                                     # minute_vent_obs > 1 & 
                                     peak_inspiratory_pressure_obs > 1 & 
                                     pressure_support_set > 1, 
                                   "nippv", 
                                   rep_len(TRUE, length(device_category)), device_category)) |>
  
  mutate(device_category = fcase(is.na(device_category) & 
                                     lead(device_category == "nippv") & 
                                     # minute_vent_obs > 1 & 
                                     peak_inspiratory_pressure_obs > 1 & 
                                     pressure_support_set > 1, 
                                   "nippv", 
                                   rep_len(TRUE, length(device_category)), device_category)) |>
  
  

  # there are times when its clearly back to CMV (resp set and volume is set but no one puts a mode back in... just leaves it blank)
  # this is usually after pressure support ... we need to classify this now as CMV. 
  # only exception to this should be when it says trach
  # There are also some without device_cat or name and they have all the variables... these should be changed too 
  mutate(
    device_category = 
      fcase(
        is.na(device_category) & 
          !str_detect(device_name, "trach") &
          tidal_volume_set > 0 & 
          resp_rate_set > 0,
        "imv",
        rep_len(TRUE, length(device_category)), device_category),
    device_name = 
      fcase(
        is.na(device_name) & 
          !str_detect(device_name, "trach") &
          tidal_volume_set > 0 & 
          resp_rate_set > 0,
        "mechanical ventilator",
        rep_len(TRUE, length(device_name)), device_name),
    mode_category = 
      fcase(
        is.na(mode_category) & 
          !str_detect(device_name, "trach") &
          tidal_volume_set > 0 & 
          resp_rate_set > 0,
        "assist control-volume control",
        rep_len(TRUE, length(mode_category)), mode_category),
    mode_name = 
      fcase(
        is.na(mode_name) & 
          !str_detect(device_name, "trach") &
          tidal_volume_set > 0 & 
          resp_rate_set > 0,
        "cmv/ac",
        rep_len(TRUE, length(mode_name)), mode_name)
  ) |> 
  
  
  # when there are duplicate times...
  group_by(clif_hospitalizations_joined_id, recorded_dttm) |> 
  
  # when bipap is part of a duplicate we need to get rid of it... 
  #     its usually when a vent is STARTED and device is carried over but it goes to a new line with lots of NAs
  #     the NA line above has the vent settings.  Its best to just drop the nippv line when its a duplicate
  #     if we don't do this... the vent settings get sent backwards across all bipap

  mutate(n = n()) |>  
  filter(
    #  essentially this is... DROP if n>1 and device_cat == nippv
    !(n > 1 & device_category == "nippv")) |> 
  
  # redo n so we keep vent settings from above... now NAs are bad around other things and we should just drop
  mutate(n = n()) |> 
  filter(
    #  essentially this is... DROP if n>1 and device_cat == NA
    !(n > 1 & is.na(device_category))) |> 
  

  
  # filter if missing everything  
  filter(
    #  essentially this is... DROP if everything missing
    !(is.na(device_category) & 
        is.na(device_name) &
        is.na(mode_category) &
        is.na(mode_name) & 
        is.na(fio2_set) &        # keeps informative fio2_set data around
        is.na(tidal_volume_set)    # keeps vent data around... this happens sort of often
      )) |> 
    
    

  # dropping duplicates for everything else but just taking the first one
  #       ffirst works WAY faster than fill up and down and slicing(1)
  ffirst() |> 
  ungroup() |> # technically don't need this  
    
  ###########
  # TEMP STOP #
  ungroup()
  ###########
  ###########

ni_toc()

```

## Resp support 

```{r}

ni_tic()

df_resp_support <- df_resp_support_1 |> 
  
# bring in hour sequences
  bind_rows(hour_sequence) |> 


#~~~~~~~~~~~~~~~~
##~~ Filling in data based on a waterfall of categories to ensure accuracy
#~~~~~~~~~~~~~~~~
  # organizing
  arrange(clif_hospitalizations_joined_id, recorded_dttm) |> 
  relocate(clif_hospitalizations_joined_id, recorded_dttm, recorded_date, recorded_hour) |> 
  
  # fill forward device category
  group_by(clif_hospitalizations_joined_id) |> 
  arrange(clif_hospitalizations_joined_id, recorded_dttm) |> 
  fill(device_category) |>
  ungroup() |> 
  
  # Record a new device_category when either (a) a new encounter, or (b) preceded by a...   
  # different device category
  mutate(
    # need to have NA as something so it gets an ID
    device_cat_f = fcase(is.na(device_category), "missing", rep_len(TRUE, length(device_category)), device_category), # cant have anything with NAs when factoring
    device_cat_f = as.integer(as.factor(device_cat_f)), # need an integer for this
    
    # getting IDs
    device_cat_id = fcumsum((
      clif_hospitalizations_joined_id != flag(clif_hospitalizations_joined_id, fill = TRUE) |           # (a)
        device_cat_f  != flag(device_cat_f, fill = TRUE)))) |>       # (b)
  
  relocate(device_cat_id, .after = recorded_hour) |> 
  
  # fill device name
  #         changed some failsafes above 4/2024 so its ok to do downup with this now
  group_by(clif_hospitalizations_joined_id, device_cat_id) |> 
  arrange(clif_hospitalizations_joined_id, recorded_dttm) |> 
  fill(device_name, .direction = "downup") |> 
  ungroup() |>
  
  
  # Record a new device_id when either (a) a new encounter, or 
  #                                    (b) preceded by a different device name.
  mutate(
    # need to have NA as something so it gets an ID
    device_name_f = fifelse(is.na(device_name), "missing", device_name), # cant have anything with NAs when factoring
    device_name_f = as.integer(as.factor(device_name_f)), # need an integer for this
    
    # getting IDs
    device_id = fcumsum((
      clif_hospitalizations_joined_id    != flag(clif_hospitalizations_joined_id, fill = TRUE) |           # (a)
        device_name_f != flag(device_name_f, fill = TRUE)))) |>      # (b)
  
  relocate(device_id, .after = recorded_hour) |> 
  
  # fill mode_category (downup)
  # there are PST that are being carried over to days before when ppl get REINTUBATED
  group_by(clif_hospitalizations_joined_id, device_id) |> 
  arrange(clif_hospitalizations_joined_id, recorded_dttm) |> 
  fill(mode_category, .direction = "downup") |> 
  ungroup() |> 
  
  # Create mode_id
  mutate(
    mode_cat_f = fifelse(is.na(mode_category), "missing", mode_category), # cant have anything with NAs when factoring
    mode_cat_f = as.integer(as.factor(mode_cat_f)), # need an integer for this
    
    mode_cat_id = fcumsum((
      device_id     != flag(device_id, fill = TRUE) |        # (a)
        mode_cat_f  != flag(mode_cat_f, fill = TRUE)))) |>   # (b)
  
  
  relocate(mode_cat_id, .after = recorded_hour) |> 
  
  # fill mode name (downup) 
  group_by(clif_hospitalizations_joined_id, mode_cat_id) |> 
  arrange(clif_hospitalizations_joined_id, recorded_dttm) |> 
  fill(mode_name, .direction = "downup") |> 
  ungroup() |>  
  
  # Create mode name id
  mutate(
    mode_name_f = fifelse(is.na(mode_name), "missing", mode_name), # cant have anything with NAs when factoring
    mode_name_f = as.integer(as.factor(mode_name_f)), # need an integer for this
    
    mode_name_id = fcumsum((
      mode_cat_id != flag(mode_cat_id, fill = TRUE) |               # (a)
        mode_name_f != flag(mode_name_f, fill = TRUE)))) |>         # (b)
  
  relocate(mode_name_id, .after = recorded_hour) |> 
  
  
  # changing fio2_set to 0.21 if room air as category
  mutate(fio2_set = if_else(is.na(fio2_set) & device_category == "room air", 21, fio2_set)) |> 
  
  # erroneous set volumes are in places where they shouldn't be for PS and trach_dome
  mutate(
    tidal_volume_set = fifelse(
      (
        mode_category == "pressure support/cpap" &    # needs to be PS/CPAP
          !is.na(pressure_support_set)                    # needs to have a PS level
      ) |
        (
          is.na(mode_category) &                      # mode cat needs to be NA
            str_detect(device_name, "trach")          # only when trach stuff
        ) |
        (
          mode_category == "pressure support/cpap" &  # needs to be PS/CPAP
            str_detect(device_name, "trach")          # only when trach stuff
        ),
      NA_integer_,
      tidal_volume_set),
    
    
  ) |>
  
  # there are ppl with t-piece that should be blow_by
  mutate(mode_category = fifelse(
    (is.na(mode_category) & 
       str_detect(device_name, "t-piece")),
    "blow by",
    mode_category
  )) |> 
  
  # carry forward the rest
  group_by(clif_hospitalizations_joined_id, mode_name_id) |>  # mode_name_id is the most granular, can go up and down
  arrange(clif_hospitalizations_joined_id, recorded_dttm) |> 
  
  # took trach out of this so we don't fill back up 3/2024
  fill(c(fio2_set, lpm_set, peep_set, tidal_volume_set, pressure_support_set, resp_rate_set, 
         # tidal_volume_obs, 
         peak_inspiratory_pressure_obs,  
         # minute_vent_obs, 
         hospital_id, location_name, location_category, in_dttm, out_dttm
  ), .direction = "downup"
  ) |>
  
  # fill trach... only down
  fill(c(tracheostomy), .direction = "down") |>
  ungroup() |> 
  
  # need to get rid of duplicates 
  distinct() |> 
  dplyr::select(
    clif_hospitalizations_joined_id,
    recorded_dttm,
    recorded_date,
    recorded_hour,
    mode_name_id,
    device_category,
    device_name,
    mode_category,
    mode_name,
    mode_cat_id,
    device_id,
    device_cat_id,
    fio2_set,
    lpm_set,
    peep_set,
    tracheostomy,
    tidal_volume_set,
    pressure_support_set,
    resp_rate_set,
    starts_with("tidal_volume_obs"),
    starts_with("mean_airway_pressure_obs"),
    peak_inspiratory_pressure_obs,
    plateau_pressure_obs,
    # obs_resp_rate,
    # minute_vent_obs,
    hospital_id,
    location_name,
    location_category,
    # in_dttm,
    # out_dttm,
    # device_cat_f,
    # device_name_f,
    # mode_cat_f,
    # mode_name_f,
  ) 

ni_toc()
```


## Vitals / ibw / labs

```{r}

######################
##~~ clean vitals and get ibw
######################

# # quickly get back to things without doing the above
# if(!exists("df_resp_support")){
#   load("output/intermediate/most_recent_save_resp.Rdata")
# 
#   df_resp_support <- read_parquet(paste0("output/intermediate/df_resp_support_", most_recent_save_resp,".parquet"))
# }

##~~~~~~~~~~~~~
##~~ spo2
##~~~~~~~~~~~~~

df_spo2 <- clif_vitals |> 
  
  # getting spo2
  filter(vital_category == "spo2") |> 
  
  # making those columns
  pivot_wider(
    id_cols = c(clif_hospitalizations_joined_id, recorded_dttm),
    names_from = vital_category,
    values_from = vital_value,
    values_fn = function(x) mean(x, na.rm = TRUE)
  ) |> 
  
  # getting date so we can group by this
  mutate(recorded_date = date(recorded_dttm),
         recorded_hour = hour(recorded_dttm)) |> 
  
  # taking the min spo2 by  DAY and HOUR and ENCOUNTER 
  group_by(clif_hospitalizations_joined_id, recorded_date, recorded_hour) |> 
  summarise(spo2 = fmin(spo2, na.rm = TRUE)) |> 
  ungroup() |> 
  dplyr::select(clif_hospitalizations_joined_id, recorded_date, recorded_hour, spo2)
  

##~~~~~~~~~~~~~
##~~ po2_arterial / so2_arterial
##~~~~~~~~~~~~~

df_po2_arterial <-  clif_labs |> 
    filter(lab_category %in% c("po2_arterial",
                             "so2_arterial")) |> 
  # going to use collected_time and not result time because its important when doing PF ratios
  # also important because we are doing daily LAPS2, so its key is when its collected
  dplyr::select(clif_hospitalizations_joined_id, recorded_dttm = lab_collect_dttm, lab_category, lab_value_numeric) |>

  # getting po2_arterial / so2_arterial
  pivot_wider(
    id_cols = c(clif_hospitalizations_joined_id, recorded_dttm),
    names_from = lab_category,
    values_from = lab_value_numeric,
    values_fn = function(x) mean(x, na.rm = TRUE)
  ) |> 
  
    # getting date so we can group by this
  mutate(recorded_date = date(recorded_dttm),
         recorded_hour = hour(recorded_dttm)) |> 
  
  # grouping by day
  group_by(clif_hospitalizations_joined_id, recorded_date) |> 
  summarise(po2_arterial = fmin(po2_arterial, na.rm = TRUE),
            so2_arterial = fmin(so2_arterial, na.rm = TRUE)
            ) |> 
  ungroup() |> 
      # getting date so we can group by this
  mutate(recorded_hour = hour(recorded_date)) |> 
  dplyr::select(clif_hospitalizations_joined_id, recorded_date, recorded_hour, po2_arterial, so2_arterial)

##~~~~~~~~~~~~~
##~~ ibw
##~~~~~~~~~~~~~

df_ibw <- clif_vitals |> 
  
  # getting height and spo2
  filter(vital_category == "height_cm") |> 
  mutate(height_in = vital_value / 2.54) |> 
  # 
  # # making those columns
  # pivot_wider(
  #   id_cols = c(clif_hospitalizations_joined_id, recorded_dttm),
  #   names_from = vital_category,
  #   values_from = vital_value,
  #   values_fn = function(x) mean(x, na.rm = TRUE)
  # ) |> 
  
  arrange(clif_hospitalizations_joined_id, recorded_dttm) |> 

  group_by(clif_hospitalizations_joined_id) |> 
  
  # taking the first height DAY and ENCOUNTER
  summarise(height_in = ffirst(height_in, na.rm = TRUE)) |> 
  ungroup() |> 
  
  # get biological sex_category
  left_join(clif_patient |> 
             dplyr::select(clif_hospitalizations_joined_id, sex_category)
            ) |> 
  
  # Calculate ibw for males
  mutate(ibw = fifelse(sex_category == "male" & height_in > 56, 50 + (2.3 * (height_in - 60)), NA_real_)) |>  # 1st % of male = 60 inches
  
  # Calculate ibw for females
  mutate(ibw = fifelse(sex_category == "female" & height_in > 56, 45.5 + (2.3 * (height_in - 60)), ibw)) |>  # 1st % of female = 56 inches
  
  # Remove records with height less than 40
  # This also removes those without a height for the whole admission, which is ok... can't do LTVV without height known
  filter(height_in >= 40) |>  # assumming these are kiddos and/or amputees which means IBW will be inaccurate
  group_by(clif_hospitalizations_joined_id) |> 
  
  # using first ibw for the encounter
  mutate(ibw = ffirst(ibw, na.rm = TRUE)) |> 
  
  # getting only the variables we want
  dplyr::select(clif_hospitalizations_joined_id, ibw, sex_category) |> 
  ungroup() |> 
  distinct()


```


## Trach Variation

```{r}

   # getting trach hospital (first one with trach == 1)
trach_hospital_id <- df_resp_support |> 
  filter(tracheostomy == 1) |> 
  dplyr::select(clif_hospitalizations_joined_id, recorded_date, trach_hospital_id = hospital_id) |> 
  arrange(clif_hospitalizations_joined_id, recorded_date) |> 
  group_by(clif_hospitalizations_joined_id) |> 
  mutate(trach_hospital_id = ffirst(trach_hospital_id)) |> 
  dplyr::select(clif_hospitalizations_joined_id, trach_hospital_id) |> 
  distinct() 

  # get the first trach day by encounter only
trach_day1 <- df_resp_support |> 
  filter(tracheostomy == 1) |> 
  dplyr::select(clif_hospitalizations_joined_id, recorded_date) |> 
  arrange(clif_hospitalizations_joined_id, recorded_date) |> 
  group_by(clif_hospitalizations_joined_id) |> 
  mutate(trach_day1 = fmin(recorded_date, na.rm = TRUE)) |> 
  dplyr::select(clif_hospitalizations_joined_id, trach_day1) |> 
  distinct() 

  # get the first ventilator day by encounter only
vent_day1 <- df_resp_support |> 
  filter(device_category == "imv") |> 
  dplyr::select(clif_hospitalizations_joined_id, recorded_date) |> 
  arrange(clif_hospitalizations_joined_id, recorded_date) |> 
  group_by(clif_hospitalizations_joined_id) |> 
  mutate(vent_day1 = fmin(recorded_date, na.rm = TRUE)) |> 
  dplyr::select(clif_hospitalizations_joined_id, vent_day1) |> 
  distinct() 
  
df_trach_variation <- df_resp_support |> 
  # keep encounters that have a tracheostomy
  mutate(tracheostomy_max = fmax(tracheostomy, na.rm = TRUE),
         .by = clif_hospitalizations_joined_id) |> 
  filter(tracheostomy_max == 1) |> 
  
  # getting vent days or trach days ONLY
  filter(device_category == "imv" | tracheostomy == 1) |> 
  
  dplyr::select(clif_hospitalizations_joined_id, recorded_date, tracheostomy) |> 
  distinct() |> 
  
  arrange(clif_hospitalizations_joined_id, recorded_date) |> 
  
  # getting vent day count
  ungroup() |> 
  
  # getting vent day count because some ppl might have days OFF the vent in between if we just do min(vent_day)
  group_by(clif_hospitalizations_joined_id) |> 
  
  # this will give 1, 2, 3, 4, 5 as long as its the same clif_hospitalizations_joined_id and trach is NA
  # once trach == 1 it parks at that max number so we can just take max
  # also... we can drop those with vent_day_count == 0 because that means they had the trach on vent day 1!!!
  mutate(vent_day_count = fcumsum(
    is.na(tracheostomy)
  )) |> 
  left_join(trach_day1) |> 
  left_join(vent_day1) |> 
  left_join(trach_hospital_id) |> 
  
  # get rid of those with zero vent_day_count (aka they were trached the whole time)
  filter(vent_day_count > 0) |> 
  dplyr::select(clif_hospitalizations_joined_id, trach_hospital_id, trach_day1, vent_day1, vent_day_count) |> 
  mutate(vent_day_count_max = max(vent_day_count, na.rm = TRUE)) |> 
  
  # true "trach_day" is vent_day_count_max - 1 because we didn't have a vent_day_zero... we started at 1
  mutate(vent_days_before_trach = vent_day_count_max - 1) |> 
  
  mutate(
  trach_day = difftime(trach_day1, vent_day1, units = "days")
    ) |>
  dplyr::select(clif_hospitalizations_joined_id, trach_hospital_id, vent_days_before_trach) |> 
  ungroup() |>
  distinct()

df_trach_variation |> glimpse()



#### should we get earliest trach and DROP everything after?
  
```


## CLIF hourly level data

```{r}

##~~~~~~~~~~~~~
##~~ Hourly Data
##~~~~~~~~~~~~~

df_hourly_resp_support1 <- df_resp_support |> 
  mutate(recorded_month = month(recorded_dttm),
         recorded_year = year(recorded_dttm)) |> 
  
  #~~ use recorded time to arrange these in order
  arrange(clif_hospitalizations_joined_id, recorded_dttm) |> 
  
  # change recorded_dttm to be the same thing at the end of the hour so we can do calculations later
  group_by(clif_hospitalizations_joined_id, recorded_date, recorded_hour) |> 
  mutate(recorded_dttm_hour_end = flast(recorded_dttm)) |> 
  ungroup() |> 
  
  #~~ vent sequence variable
      # only go up in number if (a) new category & (b) its a vent category
      # could add a gap to this as well for like hours and such
  arrange(clif_hospitalizations_joined_id, recorded_date) |> 
  mutate(gap_vent_hours = fifelse(
    device_category == "imv" | tracheostomy == 1,
    as.numeric(difftime(lead(recorded_dttm_hour_end), recorded_dttm_hour_end, units = "hours")),
    NA_real_
  )) |> 
  
  group_by(clif_hospitalizations_joined_id) |> 
  mutate(
    device_category_f = fifelse(is.na(device_category), "missing", device_category), # cant have anything with NAs when factoring
    device_category_f = as.integer(as.factor(device_category_f)), # need an integer for this
    
    # you step to the next vent sequence if
    #    1) its a vent
    #    2) its a NEW device cat
    #    3) trach is na
    #    4) gap is <= 6 between current and LEAD vent hours
    vent_episode_id = fcumsum(
      device_category == "imv" & 
        flag(device_category_f, fill = TRUE) != device_category_f &
        is.na(tracheostomy) & 
        gap_vent_hours <= 6    # this deals with the difference between the lead and the current vent_hours
      )) |> 
  ungroup() 

df_hourly_resp_support2 <- df_hourly_resp_support1 |> 
  
  # currently those with trach right away are zeros for vent_id
  # those without vent or were subsequent after vent are currently the same number as the vent before
  mutate(
    # change those not with trach/vent to NA
    vent_episode_id = fifelse(
      device_category == "imv" | tracheostomy == 1,
      vent_episode_id,
      NA_real_
    ),
    
    # we could also drop these patients if we wanted to 
    # change those trach'd on day1 without vent beforehand to enc_sequence 1
    vent_episode_id = fifelse(
      tracheostomy == 1 & vent_episode_id == 0,
      1,
      vent_episode_id
    )
  ) |> 
  
  # getting vent hour per sequence
  group_by(clif_hospitalizations_joined_id, vent_episode_id) |> 
  mutate(vent_episode_hour_seq = fcumsum(
    recorded_dttm_hour_end != flag(recorded_dttm_hour_end, fill = TRUE))
  ) |>
  # fix those not vent back to NA
  mutate(vent_episode_hour_seq = fifelse(is.na(vent_episode_id), NA_real_, vent_episode_hour_seq)) |> 
  
  
  # get max hours per vent sequence... want to drop those <=24
  mutate(vent_episode_duration_hours = fmax(vent_episode_hour_seq, na.rm = TRUE)) |>
  
  ungroup() |> 
  
  # getting vent hour per sequence per hospital
  #   there are times when that episode crosses hospitals
  #   for hospital summary we need duration split up by hospital
  group_by(clif_hospitalizations_joined_id, vent_episode_id, hospital_id) |> 
  mutate(vent_episode_hospital_hour_seq = fcumsum(
    recorded_dttm_hour_end != flag(recorded_dttm_hour_end, fill = TRUE))
  ) |> 
  
  
  # fix those not vent back to NA
  mutate(vent_episode_hospital_hour_seq = fifelse(is.na(vent_episode_id), NA_real_, vent_episode_hospital_hour_seq)) |> 
  
  
  # get max hours per vent sequence... want to drop those <=24
  mutate(vent_episode_hospital_duration_hours = fmax(vent_episode_hospital_hour_seq, na.rm = TRUE)) |>
  
  ungroup() |> 
  

  # final variables
  dplyr::select(
  clif_hospitalizations_joined_id,
  recorded_date,
  recorded_hour,
  recorded_dttm_hour_end,
  vent_episode_id,
  vent_episode_hour_seq,
  vent_episode_duration_hours,
  vent_episode_hospital_hour_seq,
  vent_episode_hospital_duration_hours,
  # recorded_dttm,
  device_category,
  device_name,
  mode_category,
  mode_name,
  fio2_set,
  recorded_month,
  recorded_year,
  peep_set,
  tracheostomy,
  tidal_volume_set,
  hospital_id,
  location_name,
  location_category,
  pressure_support_set,
  resp_rate_set,
  starts_with("tidal_volume_obs"),
  starts_with("mean_airway_pressure_obs"),
  peak_inspiratory_pressure_obs,
  plateau_pressure_obs,
  # obs_resp_rate,
  # minute_vent_obs
  ) |>
  
  #~~ group by hour so that we can take the first time point to get our information
  # this will only keep the first row if there are duplicates
  distinct(clif_hospitalizations_joined_id, recorded_date, recorded_hour, .keep_all = TRUE) |> 
  left_join(df_ibw) |> 
  left_join(df_spo2) |> 
  
  # those without ibw need to be dropped... they won't have biological sex either (this drops whole encounters)
  filter(!is.na(ibw)) |> 
  
  group_by(clif_hospitalizations_joined_id) |>
  arrange(clif_hospitalizations_joined_id, recorded_dttm_hour_end) |> 

  
  # getting each hour in sequence of their hospital stay
  mutate(enc_hour_seq = fcumsum(recorded_dttm_hour_end != flag(recorded_dttm_hour_end, fill = TRUE))) |>
  
  # getting the CALENDAR sequence... purely based on date... so day 1 can be 3 hours
  mutate(enc_calendar_date_seq = fcumsum(recorded_date != flag(recorded_date, fill = TRUE))) |>
  
  # getting DAY sequence based on which hour it is... so everyone will be 24 hours in for day 1 etc
  # Will mainly use this metric to use
  mutate(enc_day_seq = ceiling(enc_hour_seq/24)) |> 
  ungroup() |> 
  
  # merge trach variation data
  left_join(df_trach_variation) |> 
  filter(!is.na(hospital_id))


ni_toc()
  
df_hourly_resp_support <- df_hourly_resp_support2 |> 
   # final variables
    dplyr::select(
      clif_hospitalizations_joined_id,
      recorded_date,
      recorded_hour,
      # recorded_dttm_hour_end,
      vent_episode_id,
      vent_episode_hour_seq,
      vent_episode_duration_hours,
      vent_episode_hospital_hour_seq,
      vent_episode_hospital_duration_hours,
      device_category,
      device_name,
      mode_category,
      mode_name,
      fio2_set,
      recorded_month,
      recorded_year,
      peep_set,
      pressure_support_set,
      tracheostomy,
      tidal_volume_set,
      hospital_id,
      location_name,
      location_category,
      ibw,
      sex_category,
      spo2,
      enc_hour_seq,
      enc_calendar_date_seq,
      enc_day_seq,
      trach_hospital_id,
      vent_days_before_trach
      )


# duplicate check
# df_hourly_resp_support |> group_by(pick(colnames(df_hourly_resp_support))) |> filter(n() > 1) 


# 
# 
# write_parquet(df_hourly_resp_support, paste0("output/intermediate/df_hourly_resp_support_", Sys.Date(), ".parquet"))
# 
# # saving system date today
# most_recent_save_hourly <- Sys.Date()
# save(most_recent_save_hourly, file = "output/intermediate/most_recent_save_hourly.Rdata")


hosp_id_count <- df_hourly_resp_support |> count(hospital_id) |> nrow()



```





# Laps2 Score


## encounter data

```{r}
### change to hospital_icu
clif_adt_hospital <-  clif_adt |>
  mutate(location_category = factor(location_category,
                                    levels = (c("icu", "stepdown", "ward", "ed", "procedural", "other")),                                    ordered = TRUE)) |> 
  group_by(clif_hospitalizations_joined_id) |> 
  arrange(clif_hospitalizations_joined_id, location_category, in_dttm) |> 
  
  # first one will be ICU
  slice_head(n=1) |> 
  ungroup() |> 
  dplyr::select(clif_hospitalizations_joined_id,
         hospital = hospital_id)


clif_adt_admit <-  clif_adt |>
  group_by(clif_hospitalizations_joined_id) |> 
  arrange(clif_hospitalizations_joined_id, in_dttm) |> 
  slice_head(n=1) |> 
  ungroup() |> 
  dplyr::select(clif_hospitalizations_joined_id,
         first_location_category = location_category)
  

# fixing clif_labs to recorded_dttm
if ("lab_collect_dttm" %in% colnames(clif_labs)) {
  clif_labs <- clif_labs |> rename(recorded_dttm = lab_collect_dttm)
}

df_encounter_laps2_temp1 <- clif_hospitalization |> 
  left_join(clif_patient) |> 
  left_join(clif_adt_hospital) |> 
  left_join(clif_adt_admit) |> 
  
  # drop this so we can call distinct and get rid of duplicates
  dplyr::select(-hospitalization_id) |> 
  distinct() |> 
  
  # make dead/hospice column
  mutate(
    death_or_hospice_01 = fcase(
      discharge_category %in% 
        c("hospice", "dead", "expired", "died")       , 1,
      default                                         = 0
    ),
    
    death_or_hospice_01       = as.factor(death_or_hospice_01),
    
    # make 24 hours after admission variable (to collect data from first 24 hours)
    dt_24hours_after_admit    = admission_dttm + hours(24),
    
    # ED binary variable
    ed_admit_01               = fifelse(first_location_category == "ed", 1, 0, 0),
          # I think we used categorical in Rachel Kohn paper... Pat uses ED y/n
          # admission source OR you can use ED variable
    # Admit source variable (use one or the other)
    # admission_type_category      = as.factor(admission_type_category),  don't have this yet
    
    # sex_category
    female_01                 = fifelse(str_detect(sex_category, "female"), 1, 0, 0),
    
  )


```

## Anion Gap with 1 hour tolerance

```{r}

if(any(clif_labs$lab_category == "anion_gap", na.rm = TRUE)){
  
  print("you already have anion gap")
  message("we are going to remove anion_gap and recreate it for you")
  message("this way... all our analyses are the same")
  
  clif_labs <- clif_labs |> 
    filter(lab_category != "anion_gap")

    print("Anion gap removed... please proceed!!")
  
} else {
  
  print("no anion_gap found... we will make one for you!!")
  
  
}


if(any(clif_labs$lab_category == "troponin_t")){
  
  print("you already have trop_t")
  message("we are going to change to trop_i levels")
  
  clif_labs <- clif_labs |> 
    
    mutate(lab_value_numeric = fcase(
      lab_category == "troponin_t",
      lab_value_numeric / 1000,
      default = lab_value_numeric
    )) |> 
    
    mutate(lab_category = fcase(
      lab_category == "troponin_t",
      "troponin_i",
      rep_len(TRUE, length(lab_category)), lab_category
    )) 
    
    print("Troponin fixed ... please proceed!!")
  
} else {
  
  print("no need to fix troponin !!")
  
}

# Set a time tolerance, e.g., 1 hour
time_tolerance <- hours(1)

# Create recorded_date from recorded_dttm
clif_labs <- clif_labs %>%
  mutate(recorded_date = as.Date(recorded_dttm))

# Separate the lab values for sodium, bicarbonate, and chloride, and create recorded_date
sodium_df <- clif_labs %>%
  filter(lab_category == "sodium") %>%
  dplyr::select(clif_hospitalizations_joined_id, recorded_dttm, recorded_date, sodium = lab_value_numeric,
         lab_order_dttm, lab_result_dttm)

bicarbonate_df <- clif_labs %>%
  filter(lab_category == "bicarbonate") %>%
  dplyr::select(clif_hospitalizations_joined_id, recorded_dttm, recorded_date, bicarbonate = lab_value_numeric)

chloride_df <- clif_labs %>%
  filter(lab_category == "chloride") %>%
  dplyr::select(clif_hospitalizations_joined_id, recorded_dttm, recorded_date, chloride = lab_value_numeric)

# Perform a join by clif_hospitalizations_joined_id and recorded_date to limit the join size
anion_gap_df <- sodium_df %>%
  full_join(bicarbonate_df, by = c("clif_hospitalizations_joined_id", "recorded_date")) %>%
  full_join(chloride_df, by = c("clif_hospitalizations_joined_id", "recorded_date")) %>%
  
  # Calculate the time differences between recorded_dttm values
  mutate(time_diff_sodium_bicarbonate = abs(difftime(recorded_dttm.x, recorded_dttm.y, units = "mins")),
         time_diff_sodium_chloride = abs(difftime(recorded_dttm.x, recorded_dttm, units = "mins")),
         
         # Apply the time tolerance (1 hour)
         within_tolerance_bicarbonate = time_diff_sodium_bicarbonate <= time_tolerance,
         within_tolerance_chloride = time_diff_sodium_chloride <= time_tolerance) %>%
  
  # Filter to retain rows where the time differences are within the tolerance
  filter(within_tolerance_bicarbonate & within_tolerance_chloride) %>%
  
  # Calculate the anion gap: sodium - (bicarbonate + chloride)
  mutate(anion_gap = sodium - (bicarbonate + chloride)) %>%
  
  # Select the relevant columns
  dplyr::select(clif_hospitalizations_joined_id, recorded_dttm.x, anion_gap, lab_order_dttm, lab_result_dttm, recorded_date)

# Rename recorded_dttm.x to recorded_dttm for consistency
anion_gap_df <- anion_gap_df %>%
  mutate(reference_unit = "meq/l") |> 
  mutate(lab_name = "anion gap") |> 
  mutate(lab_category = "anion_gap") |>
  rename(lab_value_numeric = anion_gap) |> 
  rename(recorded_dttm = recorded_dttm.x)


clif_labs <- clif_labs |> 
              bind_rows(anion_gap_df)




```


## prelaps

```{r}

############
# Pre-Laps... want the most recent within 24 hours of admission
############



df_pre_laps2_prep1 <- 
  df_encounter_laps2_temp1 |> 
  left_join(clif_labs |> 
              filter(lab_category %in% c("anion_gap",
                                         "bicarbonate", # FROM BMP
                                         "bun",
                                         "creatinine",
                                         "sodium")),
            
            # join only if between the first 24 hours
            by = join_by(clif_hospitalizations_joined_id, 
                         dt_24hours_after_admit >= recorded_dttm) 
  ) |> 
  # Min/Max
  #     bun            max
  #     creatinine     min
  #     bun/cr         max
  #     bun/cr         most recent (prelaps)
  #     anion_gap      max
  #     bicarbonate min
  #     ag/bicarb      most recent (prelaps)
  #     sodium         min
  #     sodium         most recent (prelaps)
       
  group_by(clif_hospitalizations_joined_id, recorded_dttm, lab_category) |> 
  # using obs lets you get away with duplicates when going wider and can fix later!
  mutate(obs = row_number()) |> 
  ungroup() |> 
  pivot_wider(
    id_cols = c(clif_hospitalizations_joined_id, recorded_dttm, obs),
    names_from = lab_category,
    values_from = lab_value_numeric,
    # names_prefix = "lab_wide_",
    # values_fill = NA
  ) |>
  
  # Quick min/max for duplicates at the SAME time
  dplyr::select(-obs) |> 
  distinct() 

# find the duplicates by recorded_dttm
df_pre_laps2_prep2 <- df_pre_laps2_prep1 |> 
  arrange(clif_hospitalizations_joined_id, recorded_dttm) |> 
  group_by(clif_hospitalizations_joined_id, recorded_dttm) |>
  # mutate(n = n()) |> 
  # filter(n > 1) |> 
  summarise(
    bun               = fmax(bun, na.rm = TRUE),
    creatinine        = fmin(creatinine, na.rm = TRUE),
    bicarbonate       = fmin(bicarbonate, na.rm = TRUE),
    anion_gap         = fmax(anion_gap, na.rm = TRUE),
    sodium            = fmin(sodium, na.rm = TRUE),

    .groups = "drop"
  ) |> 
  
  # replace inf with NAs
  mutate(across(where(is.numeric) & !c(clif_hospitalizations_joined_id), ~ na_if(.x, is.infinite(.x)))) |> 
  
  # get rid of duplicate rows
  distinct() |> 

# df_pre_laps2_prep2 <- df_pre_laps2_prep1 |> 
#   anti_join(df_pre_laps2_duplicates |> 
#               dplyr::select(clif_hospitalizations_joined_id, recorded_dttm),
#             join_by(clif_hospitalizations_joined_id, recorded_dttm)) |> 
#   bind_rows(df_pre_laps2_duplicates) |> 
  
  # order appropriately for when getting last labs for "most recent"
  arrange(clif_hospitalizations_joined_id, recorded_dttm) |> 
  
  # carry forward
  group_by(clif_hospitalizations_joined_id) |> 
  
  # fill in ... its only the first 24 hours so its ok
  fill(c(bun,
         creatinine,
         bicarbonate,
         anion_gap,
         sodium), 
       .direction = "downup") |>
  
  # ? time shift if needed
  
  # get the most recent lab (aka last) based on recorded_dttm
    mutate(
      bun_recent                = flast(bun, na.rm = TRUE),
      creatinine_recent         = flast(creatinine, na.rm = TRUE),
      bicarbonate_recent        = flast(bicarbonate, na.rm = TRUE),
      anion_gap_recent          = flast(anion_gap, na.rm = TRUE),
      sodium_recent             = flast(sodium, na.rm = TRUE)
           ) |> 
  ungroup() |> 
  dplyr::select(clif_hospitalizations_joined_id, ends_with("recent")) |> 

  # final pre-laps labs needed
  mutate(
    # make ratios
    sodium              = sodium_recent,
    bun_cr_ratio        = bun_recent/creatinine_recent,
    ag_hco3_ratio       = (anion_gap_recent/bicarbonate_recent)*1000
  ) |> 
  dplyr::select(-ends_with("recent")) |> 
  
  # replace inf with NAs (sometimes it happens when divided by zero)
  mutate(across(where(is.numeric) & !c(clif_hospitalizations_joined_id), ~ na_if(.x, is.infinite(.x)))) |> 
  
  distinct()
  



# replacing all the admission values that are missing with normal values
        # replace BUNCREAT_RECENT_24=8 if BUNCREAT_RECENT_24==. // 4724
        # replace NA_RECENT_24=135 if NA_RECENT_24==. // 4270
        # replace AGHCO3_RECENT_24=200 if AGHCO3_RECENT_24==. // 5103
        
 df_pre_laps_merging <- df_pre_laps2_prep2 |> 
   left_join(
     df_encounter_laps2_temp1 |> 
       dplyr::select(
         clif_hospitalizations_joined_id,
         admission_dttm,
         age = age_at_admission,
         female_01,
         ed_admit_01,
         # admission_type_category,
         death_or_hospice_01
       ) |> 
       distinct(),
     join_by(clif_hospitalizations_joined_id)
   ) |> 
   mutate(
    age_cat = case_when(
      age >= 85 ~ 5,
      age >= 75 ~ 4,
      age >= 65 ~ 3,
      age >= 40 ~ 2,
      TRUE ~ 1
    )
  ) |> 
   mutate(
      # impute to normal (8)
    bun_cr_ratio = fifelse(is.na(bun_cr_ratio), 8, bun_cr_ratio),
     
    buncreat_cat = case_when(
      bun_cr_ratio        >= 24 ~ 4,
      bun_cr_ratio        >= 16 ~ 3,
      bun_cr_ratio        < 8   ~ 2,
      TRUE                      ~ 1
    )
  ) |> 
     mutate(
      # impute to normal (135)
    sodium = fifelse(is.na(sodium), 135, sodium),
    na_cat = case_when(
      sodium            >= 155  ~ 7,
      sodium            >= 149  ~ 6,
      sodium            >= 146  ~ 5,
      sodium            >= 132  ~ 4,
      sodium            >= 129  ~ 3,
      sodium            < 129   ~ 2,
      TRUE                      ~ 1
    )
  ) |> 
     mutate(
      # impute to normal (200)
    ag_hco3_ratio = fifelse(is.na(ag_hco3_ratio), 200, ag_hco3_ratio),
    aghco3_cat = case_when(
      ag_hco3_ratio        >= 600  ~ 4,
      ag_hco3_ratio        >= 400  ~ 3,
      ag_hco3_ratio        < 200   ~ 2,
      TRUE                         ~ 1,
    )
  )
 
 # logistic regression to obtain p(death/hospice)
 lr_model <-
  glm(
    death_or_hospice_01 ~
      age +
      female_01 +
      ed_admit_01 +
      bun_cr_ratio +
      sodium +
      ag_hco3_ratio,
    family = "binomial",
    data   = df_pre_laps_merging
  )

tidy(lr_model)
pR2(lr_model)["McFadden"]

df_pre_laps_final  <- df_pre_laps_merging |> 
  mutate(p_death = predict(lr_model, df_pre_laps_merging, type = "response")) |> 
  mutate(
    high_risk = case_when(
      p_death >= 0.06 ~ 1,
      TRUE ~ 0
    )) %>%
  dplyr::select(clif_hospitalizations_joined_id, p_death, high_risk)

```


The above has like 58% high risk... that seems too high
See testing below but ultimately I think its ok to use the predictions as above and not the published ones below.  
R2 is 17 (using GLM) vs 11 (using numbers from Escobar study) when doing glm(death ~ p_death) in isolation.  Despite having a high number of high risk (58% with glm vs 10% with escobar numbers), i think it will work out ok


## laps2 labs ALL

```{r}
##~~~~~~~~~~~~~~~~~~~~~
# LAPS2 time
##~~~~~~~~~~~~~~~~~~~~~
 
df_laps2_prep1 <- 
  df_encounter_laps2_temp1 |> 
  dplyr::select(clif_hospitalizations_joined_id, admission_dttm, discharge_dttm) |> 
  distinct() |> 
  # slice_head(n=1000) |> 
  left_join(clif_labs |> 
              filter(lab_category %in% c("albumin",
                                     "anion_gap",
                                     "bilirubin_total",
                                     "bun",
                                     "bicarbonate",
                                     "creatinine",
                                     "glucose_serum",
                                     # "hematocrit",
                                     "hemoglobin",
                                     "lactate",
                                     "pco2_arterial",
                                     "po2_arterial",
                                     "ph_arterial",
                                     "platelet_count",
                                     "so2_arterial",
                                     "sodium",
                                     "troponin_i",
                                     "wbc")) |> 
              dplyr::select(-lab_order_dttm),
            
            # join only if between the first 24 hours
            by = join_by(clif_hospitalizations_joined_id)) |> 
  # making all hemoglobin into hematocrits
  mutate(lab_value_numeric = fifelse(lab_category == "hemoglobin", lab_value_numeric*3, lab_value_numeric)) |> 
  mutate(lab_category = fifelse(lab_category == "hemoglobin", "hematocrit", lab_category)) |> 
  mutate(lab_category = fifelse(lab_category == "lactate", "lactic_acid", lab_category)) |> 

    # using obs lets you get away with duplicates when going wider and can fix later!
  group_by(clif_hospitalizations_joined_id, recorded_dttm, lab_category) |> 
  mutate(obs = row_number()) |> 
  ungroup() |> 
  
  # pivot wider to get columns
  pivot_wider(
    id_cols = c(clif_hospitalizations_joined_id, recorded_dttm, obs),
    names_from = lab_category,
    values_from = lab_value_numeric,
    # names_prefix = "lab_wide_",
    # values_fill = NA
  ) |>
  
  # Quick min/max for duplicates at the SAME time
  dplyr::select(-obs) |> 
  mutate(lab_date = date(recorded_dttm)) |> 
  distinct() 

ni_tic()
# find the duplicates by day
df_laps2_prep2 <- df_laps2_prep1 |>
  # arrange(clif_hospitalizations_joined_id, recorded_dttm) |>
  group_by(clif_hospitalizations_joined_id, lab_date) |>
  # mutate(n = n()) |>
  # filter(n > 1) |>
  summarise(
    albumin             = fmin(albumin, na.rm = TRUE),
    anion_gap           = fmax(anion_gap, na.rm = TRUE),
    bilirubin_total     = fmax(bilirubin_total, na.rm = TRUE),
    bun                 = fmax(bun, na.rm = TRUE),
    bicarbonate         = fmin(bicarbonate, na.rm = TRUE),
    creatinine          = fmax(creatinine, na.rm = TRUE),
    glucose_serum       = fmin(glucose_serum, na.rm = TRUE),
    hematocrit          = fmax(hematocrit, na.rm = TRUE),
    lactic_acid         = fmax(lactic_acid, na.rm = TRUE),
    pco2_arterial       = fmax(pco2_arterial, na.rm = TRUE),
    po2_arterial        = fmax(po2_arterial, na.rm = TRUE),
    ph_arterial         = fmin(ph_arterial, na.rm = TRUE),
    platelet_count      = fmin(platelet_count, na.rm = TRUE),
    so2_arterial        = fmin(so2_arterial, na.rm = TRUE),
    sodium              = fmin(sodium, na.rm = TRUE),
    troponin            = fmax(troponin_i, na.rm = TRUE),
    wbc                 = fmin(wbc, na.rm = TRUE
    ), 

    .groups = "drop"
  ) |> 
  
    # bun_cr_ratio
  mutate(
    bun_cr_ratio = bun / creatinine
  ) |> 
  
    mutate(across(where(is.numeric) & !c(clif_hospitalizations_joined_id), ~ na_if(.x, is.infinite(.x)))) |> 
  # get rid of duplicate rows
  distinct() 




# |> 
#   group_by(clif_hospitalizations_joined_id) |> 
#   # carryforward after prelaps imputation
#     fill(c(
#       albumin,
#       anion_gap,
#       bilirubin_total,
#       bun,
#       bicarbonate,
#       creatinine,
#       glucose,
#       hematocrit,
#       lactic_acid,
#       pco2_arterial,
#       po2_arterial,
#       ph_arterial,
#       platelet_count,
#       so2_arterial,
#       sodium,
#       troponin,
#       wbc,
#       ), 
#       .direction = "down") |>
  
  ni_toc()


            
```

## Labs day one

```{r}
df_laps2_dayone_prep1 <- 
  df_encounter_laps2_temp1 |> 
  dplyr::select(clif_hospitalizations_joined_id, admission_dttm, discharge_dttm, dt_24hours_after_admit) |> 
  distinct() |> 
  left_join(clif_labs |> 
              filter(lab_category %in% c("albumin",
                                     "anion_gap",
                                     "bilirubin_total",
                                     "bun",
                                     "bicarbonate",
                                     "creatinine",
                                     "glucose_serum",
                                     # "hematocrit",
                                     "hemoglobin",
                                     "lactate",
                                     "pco2_arterial",
                                     "po2_arterial",
                                     "ph_arterial",
                                     "platelet_count",
                                     "so2_arterial",
                                     "sodium",
                                     "troponin_i",
                                     "wbc")) |>            
              dplyr::select(-lab_order_dttm),
            
            # join only if between the first 24 hours
            by = join_by(clif_hospitalizations_joined_id, 
                         dt_24hours_after_admit >= recorded_dttm)
  ) |> 

  # making all hemoglobin into hematocrits
  mutate(lab_value_numeric = fifelse(lab_category == "hemoglobin", lab_value_numeric*3, lab_value_numeric)) |> 
  mutate(lab_category = fifelse(lab_category == "hemoglobin", "hematocrit", lab_category)) |> 
  mutate(lab_category = fifelse(lab_category == "lactate", "lactic_acid", lab_category)) |> 
  
  group_by(clif_hospitalizations_joined_id, recorded_dttm, lab_category) |> 
  # using obs lets you get away with duplicates when going wider and can fix later!
  mutate(obs = row_number()) |> 
  ungroup() |> 
  
  # pivot wider to get columns
  pivot_wider(
    id_cols = c(clif_hospitalizations_joined_id, recorded_dttm, obs, admission_dttm),
    names_from = lab_category,
    values_from = lab_value_numeric,
    # names_prefix = "lab_wide_",
    # values_fill = NA
  ) |>
  
  # Quick min/max for duplicates at the SAME time
  dplyr::select(-obs) |> 
  distinct() 

# get min/max
df_laps2_dayone_prep2 <- df_laps2_dayone_prep1 |>
  arrange(clif_hospitalizations_joined_id, recorded_dttm) |> 
  
  # filter out those without a recorded_dttm so we don't add a date later
  filter(!is.na(recorded_dttm)) |> 
  # if we make "lab_date" the encounter date... then we are switching all labs within 24hr of admission to the "day 1" in a sneaky fashion
  mutate(lab_date = date(admission_dttm)) |> # make it all the same day 1
  group_by(clif_hospitalizations_joined_id, lab_date) |>
  # mutate(n = n()) |>
  # filter(n > 1) |>
  summarise(
    albumin             = fmin(albumin, na.rm = TRUE),
    anion_gap           = fmax(anion_gap, na.rm = TRUE),
    bilirubin_total     = fmax(bilirubin_total, na.rm = TRUE),
    bun                 = fmax(bun, na.rm = TRUE),
    bicarbonate         = fmin(bicarbonate, na.rm = TRUE),
    creatinine          = fmax(creatinine, na.rm = TRUE),
    glucose_serum       = fmin(glucose_serum, na.rm = TRUE),
    hematocrit          = fmax(hematocrit, na.rm = TRUE),
    lactic_acid         = fmax(lactic_acid, na.rm = TRUE),
    pco2_arterial       = fmax(pco2_arterial, na.rm = TRUE),
    po2_arterial        = fmax(po2_arterial, na.rm = TRUE),
    ph_arterial         = fmin(ph_arterial, na.rm = TRUE),
    platelet_count      = fmin(platelet_count, na.rm = TRUE),
    so2_arterial        = fmin(so2_arterial, na.rm = TRUE),
    sodium              = fmin(sodium, na.rm = TRUE),
    troponin            = fmax(troponin_i, na.rm = TRUE),
    wbc                 = fmin(wbc, na.rm = TRUE
    ),

    .groups = "drop"
  ) |> 
  
    # bun_cr_ratio
  mutate(
    bun_cr_ratio = bun / creatinine
  ) |> 
  
    # replace inf with NAs
  mutate(across(where(is.numeric) & !c(clif_hospitalizations_joined_id), ~ na_if(.x, is.infinite(.x)))) |> 

    # get rid of duplicate rows
  distinct() 



# switch out the day 1 stuff 
df_laps2_final <- df_laps2_prep2 |> anti_join(df_laps2_dayone_prep2 |> 
                              dplyr::select(clif_hospitalizations_joined_id, lab_date)) |> 
  bind_rows(df_laps2_dayone_prep2) |> 
  arrange(clif_hospitalizations_joined_id, lab_date) |> 
  

  # filter out NA for lab date
  filter(!is.na(lab_date))

  
```


## GCS


```{r}



df_gcs_merge_ready <- clif_gcs |>
  mutate(vital_value = case_when(
    gcs_total %in% c(14, 15)                ~ 1,
    gcs_total %in% c(8, 9, 10, 11, 12, 13)  ~ 3,
    gcs_total %in% c(2, 3, 4, 5, 6, 7)      ~ 4,
    TRUE                                    ~ NA   # we account for this when doing the points later
  )) |>
  mutate(vital_category = "gcs") |> 
  dplyr::select(clif_hospitalizations_joined_id,
         vital_category,
         recorded_dttm,
         vital_value,
         )


df_vitals_gcs <- clif_vitals |>
              filter(vital_category %in% c("temp_c",
                                       "sbp",
                                       "spo2",
                                       "heart_rate",
                                       "respiratory_rate")) |>
              # dplyr::select(-meas_site_name) |> 
              
                          # get gcs in there now
              bind_rows(df_gcs_merge_ready)


```

## vitals all

```{r}

df_laps2_vitals_prep1 <- 
  df_encounter_laps2_temp1 |> 
  dplyr::select(clif_hospitalizations_joined_id, admission_dttm, discharge_dttm) |>
  distinct() |> 
  left_join(df_vitals_gcs,
            
            # join by the encounter ID
            by = join_by(clif_hospitalizations_joined_id)) |> 
  
    group_by(clif_hospitalizations_joined_id, recorded_dttm, vital_category) |> 
  # using obs lets you get away with duplicates when going wider and can fix later!
  mutate(obs = row_number()) |> 
  ungroup() |> 
  pivot_wider(
    id_cols = c(clif_hospitalizations_joined_id, recorded_dttm, obs),
    names_from = vital_category,
    values_from = vital_value,
    # names_prefix = "vital_wide_",
    # values_fill = NA
  ) |>
  
  #fixing temp 
  mutate(temp = (temp_c * (9/5)) + 32) |> 
  
  #renaming resp
  rename(resp = respiratory_rate) |> 
  
  # Quick min/max for duplicates at the SAME time
  dplyr::select(-obs) |> 
  distinct() 

# find the duplicates by day
df_laps2_vitals_prep2 <- df_laps2_vitals_prep1 |>
  arrange(clif_hospitalizations_joined_id, recorded_dttm) |>
  mutate(vital_date = date(recorded_dttm)) |> 
  group_by(clif_hospitalizations_joined_id, vital_date) |>
  # mutate(n = n()) |>
  # filter(n > 1) |>
  summarise(
    temp                = fmin(temp, na.rm = TRUE),
    sbp                 = fmin(sbp, na.rm = TRUE),
    spo2                = fmin(spo2, na.rm = TRUE),
    heart_rate               = fmax(heart_rate, na.rm = TRUE),
    resp                = fmax(resp, na.rm = TRUE),
    gcs                 = fmax(gcs, na.rm = TRUE),  # remember the groupings... higher is lower gcs
    # shock_index         = fmax(shock_index, na.rm = TRUE),  # do this after

    .groups = "drop"
  ) |> 
  mutate(
    shock_index = heart_rate / sbp
  ) |> 
  # replace inf with NAs
  mutate(across(where(is.numeric) & !c(clif_hospitalizations_joined_id), ~ na_if(.x, is.infinite(.x)))) |> 

  # get rid of duplicate rows
  distinct() 




```


## vitals day one

```{r}


df_laps2_vitals_dayone_prep1 <- 
  df_encounter_laps2_temp1 |> 
  dplyr::select(clif_hospitalizations_joined_id, admission_dttm, discharge_dttm, dt_24hours_after_admit) |>
  distinct() |> 
  left_join(df_vitals_gcs,
            
            # join only if between the first 24 hours
            by = join_by(clif_hospitalizations_joined_id, 
                         dt_24hours_after_admit >= recorded_dttm)
  ) |> 
  
    group_by(clif_hospitalizations_joined_id, recorded_dttm, vital_category) |> 
  # using obs lets you get away with duplicates when going wider and can fix later!
  mutate(obs = row_number()) |> 
  ungroup() |> 
  pivot_wider(
    id_cols = c(clif_hospitalizations_joined_id, recorded_dttm, obs, admission_dttm),
    names_from = vital_category,
    values_from = vital_value,
    # names_prefix = "vital_wide_",
    # values_fill = NA
  ) |>
  
  #fixing temp 
  mutate(temp = (temp_c * (9/5)) + 32) |> 
  
  #renaming resp
  rename(resp = respiratory_rate) |> 
  
  
  # Quick min/max for duplicates at the SAME time
  dplyr::select(-obs) |> 
  

  distinct() 

# find the duplicates by day
df_laps2_vitals_dayone_prep2 <- df_laps2_vitals_dayone_prep1 |>
  arrange(clif_hospitalizations_joined_id, recorded_dttm) |>
  
  # filter out those without a recorded_dttm so we don't add a date later
  filter(!is.na(recorded_dttm)) |> 
  
  # if we make "lab_date" the encounter date... then we are switching all labs within 24hr of admission to the "day 1" in a sneaky fashion
  #using encounter admit date to collect everything together
  mutate(vital_date = date(admission_dttm)) |> 
  group_by(clif_hospitalizations_joined_id, vital_date) |>
  # mutate(n = n()) |>
  # filter(n > 1) |>
  summarise(
    temp                = fmin(temp, na.rm = TRUE),
    sbp                 = fmin(sbp, na.rm = TRUE),
    spo2                = fmin(spo2, na.rm = TRUE),
    heart_rate          = fmax(heart_rate, na.rm = TRUE),
    resp                = fmax(resp, na.rm = TRUE),
    gcs                 = fmax(gcs, na.rm = TRUE),
    # shock_index         = fmax(shock_index, na.rm = TRUE),  # do this after

    .groups = "drop"
  ) |> 
  mutate(
    shock_index         = heart_rate / sbp
  ) |> 
  
  # replace inf with NAs
  mutate(across(where(is.numeric) & !c(clif_hospitalizations_joined_id), ~ na_if(.x, is.infinite(.x)))) |> 

  # get rid of duplicate rows
  distinct() 

# switch out the day 1 stuff 
df_laps2_vitals_final <- df_laps2_vitals_prep2 |> anti_join(df_laps2_vitals_dayone_prep2 |> 
                              dplyr::select(clif_hospitalizations_joined_id, vital_date)) |> 
  bind_rows(df_laps2_vitals_dayone_prep2) |> 
  arrange(clif_hospitalizations_joined_id, vital_date) |> 
  
  # drop rows with no recorded date
  filter(!is.na(vital_date))



```



## Laps2 code

```{r}

df_laps_calc <- df_laps2_final |> 
  # get same naming for date
  rename(recorded_date = lab_date) |> 
  full_join(df_laps2_vitals_final |> 
              rename(recorded_date = vital_date),
            join_by(clif_hospitalizations_joined_id, recorded_date)
            ) |> 
  full_join(df_pre_laps_final)
  
clif_laps2_scores <- df_laps_calc |> 
  group_by(clif_hospitalizations_joined_id, recorded_date) |> 
  
  mutate(laps2 = 0) |> 
  
  mutate(
    # lactic_acid and pH 
    laps2 = fcase(
      
      # missing data
      is.na(ph_arterial)  &                      is.na(lactic_acid) & high_risk == 0 , laps2 + 0,
      is.na(ph_arterial)  &                      is.na(lactic_acid) & high_risk == 1 , laps2 + 15,
      is.na(ph_arterial)  &                      lactic_acid < 2    & high_risk == 0 , laps2 + 0,
      is.na(ph_arterial)  &                      lactic_acid < 2    & high_risk == 1 , laps2 + 12,
      is.na(ph_arterial)  & lactic_acid >= 2   & lactic_acid < 4    & high_risk == 0 , laps2 + 12,
      is.na(ph_arterial)  & lactic_acid >= 2   & lactic_acid < 4    & high_risk == 1 , laps2 + 15,
      is.na(ph_arterial)  &                      lactic_acid >= 4   & high_risk == 0 , laps2 + 26,
      is.na(ph_arterial)  &                      lactic_acid >= 4   & high_risk == 1 , laps2 + 30,
      ph_arterial < 7.2   &                      is.na(lactic_acid) & high_risk == 0 , laps2 + 13,
      ph_arterial < 7.2   &                      is.na(lactic_acid) & high_risk == 1 , laps2 + 19,
      ph_arterial >= 7.2  & ph_arterial < 7.35 & is.na(lactic_acid) & high_risk == 0 , laps2 + 5,
      ph_arterial >= 7.2  & ph_arterial < 7.35 & is.na(lactic_acid) & high_risk == 1 , laps2 + 15,
      ph_arterial >= 7.35 & ph_arterial < 7.45 & is.na(lactic_acid) & high_risk == 0 , laps2 + 0,
      ph_arterial >= 7.35 & ph_arterial < 7.45 & is.na(lactic_acid) & high_risk == 1 , laps2 + 12,
      ph_arterial >= 7.45 &                      is.na(lactic_acid) & high_risk == 0 , laps2 + 12,
      ph_arterial >= 7.45 &                      is.na(lactic_acid) & high_risk == 1 , laps2 + 15,

      # complete data
      ph_arterial < 7.2   &                      lactic_acid <  2                   , laps2 + 13,
      ph_arterial < 7.2   & lactic_acid >= 2   & lactic_acid <  4                   , laps2 + 19,
      ph_arterial < 7.2   &                      lactic_acid >= 4                   , laps2 + 34,
      ph_arterial >= 7.2  & ph_arterial < 7.35 & lactic_acid <  2                   , laps2 + 5,
      ph_arterial >= 7.2  & ph_arterial < 7.35 & lactic_acid >= 2 & lactic_acid < 4 , laps2 + 15,
      ph_arterial >= 7.2  & ph_arterial < 7.35 & lactic_acid >= 4                   , laps2 + 25,
      ph_arterial >= 7.35 & ph_arterial < 7.45 & lactic_acid <  2                   , laps2 + 0,
      ph_arterial >= 7.35 & ph_arterial < 7.45 & lactic_acid >= 2 & lactic_acid < 4 , laps2 + 12,
      ph_arterial >= 7.35 & ph_arterial < 7.45 & lactic_acid >= 4                   , laps2 + 26,
      ph_arterial >= 7.45 &                      lactic_acid <  2                   , laps2 + 12,
      ph_arterial >= 7.45 & lactic_acid >= 2 &   lactic_acid <  4                   , laps2 + 15,
      ph_arterial >= 7.45 &                      lactic_acid >= 4                   , laps2 + 30,
      default                                                                       = laps2
      ),
    
    # Sodium
    laps2 = fcase(
      is.na(sodium)                                                             , laps2 + 0,
      sodium <  129                                                             , laps2 + 14,
      sodium >= 129 & sodium < 135                                              , laps2 + 7,
      sodium >= 135 & sodium < 146                                              , laps2 + 0,
      sodium >= 146                                                             , laps2 + 4,
      default                                                                   = laps2
    ),
    
    # Bilirubin
    laps2 = fcase(
      is.na(bilirubin_total)                                                    , laps2 + 0,
      bilirubin_total <  2                                                      , laps2 + 0,
      bilirubin_total >= 2 & bilirubin_total < 3                                , laps2 + 11,
      bilirubin_total >= 3 & bilirubin_total < 5                                , laps2 + 18,
      bilirubin_total >= 5 & bilirubin_total < 8                                , laps2 + 25,
      bilirubin_total >= 8                                                      , laps2 + 41,
      TRUE , laps2
    ),
    
    # BUN
    laps2 = fcase(
      is.na(bun)                                                                , laps2 + 0,
      bun < 18                                                                  , laps2 + 0,
      bun >= 18 & bun < 20                                                      , laps2 + 11,
      bun >= 20 & bun < 40                                                      , laps2 + 12,
      bun >= 40 & bun < 80                                                      , laps2 + 20,
      bun >= 80                                                                 , laps2 + 25,
      default                                                                   = laps2
    ),
    
    # Creatinine
    laps2 = fcase(
      is.na(creatinine)                                                         , laps2 + 0,
      creatinine < 1                                                            , laps2 + 0,
      creatinine >= 1 & creatinine < 2                                          , laps2 + 6,
      creatinine >= 2 & creatinine < 4                                          , laps2 + 11,
      creatinine >= 4                                                           , laps2 + 5,
      default                                                                   = laps2
    ),
    
    # BUN/Cr Ratio
    laps2 = fcase(
      is.na(bun_cr_ratio)                                                       , laps2 + 0,
      bun_cr_ratio < 25                                                         , laps2 + 0,
      bun_cr_ratio >= 25                                                        , laps2 + 10,
      default                                                                   = laps2
    ),
    
    # Albumin
    laps2 = fcase(
      is.na(albumin)                                                            , laps2 + 0,
      albumin < 2                                                               , laps2 + 31,
      albumin >= 2 & albumin < 2.5                                              , laps2 + 15,
      albumin >= 2.5                                                            , laps2 + 0,
      default                                                                   = laps2
    ),
    
    # glucose_serum
    laps2 = fcase(
      is.na(glucose_serum)                                                      , laps2 + 0,
      glucose_serum < 40                                                        , laps2 + 10,
      glucose_serum >= 40 & glucose_serum < 60                                  , laps2 + 10,
      glucose_serum >= 60 & glucose_serum < 200                                 , laps2 + 0,
      glucose_serum >= 200                                                      , laps2 + 3,
      default                                                                   = laps2
    ),
    
    # Hematocrit (Hct)
    laps2 = fcase(
      is.na(hematocrit)                                                         , laps2 + 0,
      hematocrit < 20                                                           , laps2 + 7,
      hematocrit >= 20 & hematocrit < 40                                        , laps2 + 8,
      hematocrit >= 40 & hematocrit < 50                                        , laps2 + 0,
      hematocrit >= 50                                                          , laps2 + 3,
      default                                                                   = laps2
    ),
    
    # WBC
    laps2 = fcase(
      is.na(wbc) &                                               high_risk == 0 , laps2 + 0,
      is.na(wbc) &                                               high_risk == 1 , laps2 + 32,
      wbc < 5                                                                   , laps2 + 8,
      wbc >= 5 & wbc < 13                                                       , laps2 + 0,
      wbc >= 13                                                                 , laps2 + 11,
      default                                                                   = laps2
    ),
    
  # pco2_arterial
    laps2 = fcase(
      is.na(pco2_arterial)                                                      , laps2 + 0,
      pco2_arterial < 35                                                        , laps2 + 7,
      pco2_arterial >= 35 & pco2_arterial < 45                                  , laps2 + 0,
      pco2_arterial >= 45 & pco2_arterial < 55                                  , laps2 + 11,
      pco2_arterial >= 55 & pco2_arterial < 65                                  , laps2 + 13,
      pco2_arterial >= 65                                                       , laps2 + 12,
      default                                                                   = laps2
    ),
  
  # po2_arterial
    laps2 = fcase(
      is.na(po2_arterial)                                                       , laps2 + 0,
      po2_arterial < 50                                                         , laps2 + 8,
      po2_arterial >= 50 & po2_arterial < 120                                   , laps2 + 0,
      po2_arterial >= 120                                                       , laps2 + 12,
      default                                                                   = laps2
    ),
  
  # Troponin
    laps2 = fcase(
      is.na(troponin)  &                                       high_risk == 0 , laps2 + 0,
      is.na(troponin)  &                                       high_risk == 1 , laps2 + 9,
      troponin <  0.01                                                        , laps2 + 0,
      troponin >= 0.01 & troponin < 0.2                                       , laps2 + 8,
      troponin >= 0.2  & troponin < 1                                         , laps2 + 17,
      troponin >= 1    & troponin < 3                                         , laps2 + 19,
      troponin >= 3                                                           , laps2 + 25,
      default                                                                   = laps2
    ),
  
  # Temp
    laps2 = fcase(
      is.na(temp)                                                               , laps2 + 0,
      temp < 96                                                                 , laps2 + 20,
      temp >= 96    & temp < 100.5                                              , laps2 + 0,
      temp >= 100.5                                                             , laps2 + 3,
      default                                                                   = laps2
    ),

  # heart_rate
    laps2 = fcase(
      is.na(heart_rate)                                                         , laps2 + 0,
      heart_rate < 60                                                           , laps2 + 7,
      heart_rate >= 60  & heart_rate < 110                                      , laps2 + 0,
      heart_rate >= 110 & heart_rate < 140                                      , laps2 + 7,
      heart_rate >= 140                                                         , laps2 + 10,
      default                                                                   = laps2
    ),
  
  # Resp
    laps2 = fcase(
      is.na(resp)                                                               , laps2 + 0,
      resp < 20                                                                 , laps2 + 0,
      resp >= 20 & resp < 30                                                    , laps2 + 11,
      resp >= 30                                                                , laps2 + 21,
      default                                                                   = laps2
    ),
    
  # SBP
    laps2 = fcase(
      is.na(sbp)                                                                , laps2 + 0,
      sbp < 75                                                                  , laps2 + 22,
      sbp >= 75  & sbp < 90                                                     , laps2 + 13,
      sbp >= 90  & sbp < 120                                                    , laps2 + 5,
      sbp >= 120 & sbp < 140                                                    , laps2 + 0,
      sbp >= 140 & sbp < 160                                                    , laps2 + 8,
      sbp >= 160                                                                , laps2 + 14,
      default                                                                   = laps2
    ),
  
  # Shock
    laps2 = fcase(
      is.na(shock_index)                                                        , laps2 + 0,
      shock_index < 0.65                                                        , laps2 + 0,
      shock_index >= 0.65 & shock_index < 0.85                                  , laps2 + 8,
      shock_index >= 0.85                                                       , laps2 + 17,
      default                                                                   = laps2
    ),
  
  # O2Sat
    laps2 = fcase(
      is.na(spo2) &                                              high_risk == 0 , laps2 + 0,
      is.na(spo2) &                                              high_risk == 1 , laps2 + 22,
      spo2 <  90                                                                , laps2 + 22,
      spo2 >= 90  & spo2 < 94                                                   , laps2 + 12,
      spo2 >  94                                                                , laps2 + 0,
      default                                                                   = laps2
    ),
  
  # Neuro (GCS)
    laps2 = fcase(
      is.na(gcs) &                                               high_risk == 0 , laps2 + 16,
      is.na(gcs) &                                               high_risk == 1 , laps2 + 21,
      gcs == 1                                                                  , laps2 + 0,
      gcs == 2                                                                  , laps2 + 16,
      gcs == 3                                                                  , laps2 + 21,
      gcs == 4                                                                  , laps2 + 36,
      default                                                                   = laps2
    )
  )

  


```




# Intermediate save

```{r}
# Define the function
save_data_frames <- function(data_frame_names, output_dir = "output/intermediate/clean_db") {
  # Ensure the output directory exists
  if (!dir.exists(output_dir)) {
    dir.create(output_dir, recursive = TRUE)
  }
  
  # Save the current date
  most_recent_save_resp <- Sys.Date()
  save(most_recent_save_resp, file = file.path(output_dir, "most_recent_save_resp.Rdata"))
 
  
  # Iterate over each data frame name
  for (df_name in data_frame_names) {
    # Check if the data frame exists in the global environment
    if (exists(df_name, envir = .GlobalEnv)) {
      # Get the data frame
      df <- get(df_name, envir = .GlobalEnv)
      
      # Construct the output file path
      output_file <- file.path(output_dir, paste0(df_name, "_.", Sys.Date(), ".parquet"))
      
      # Save the data frame to Parquet
      arrow::write_parquet(df, output_file)
      
      # Optional: Print a message
      message("Saved ", df_name, " to ", output_file)
    } else {
      warning("Data frame '", df_name, "' does not exist in the global environment.")
    }
  }
}

# List of data frame names
data_frames_to_save <- 
  c(
   "clif_adt",
   "clif_demographics_combined",
   "clif_gcs",
   "clif_hospitalization",
   "clif_hospitalization_start",
   "clif_labs",
   "clif_patient",
   "clif_patient_assessments",
   "clif_patient_start",
   "clif_respiratory_support",
   "clif_vitals",
   "df_cohort_keep",
   "df_hourly_resp_support",
   "df_ibw",
   "df_po2_arterial",
   "df_resp_support",
   "df_resp_support_1",
   "df_spo2",
   "df_trach_variation",
   "clif_adt_hospital",
   "clif_adt_admit",
   "clif_laps2_scores",
   "hospital_order")
 
save(data_frames_to_save, file = file.path("output/intermediate/clean_db/data_frames_to_save.Rdata"))

# Call the function
save_data_frames(data_frames_to_save)
```


```{r}

Sys.time()
end.time.full <- Sys.time()
full_time.taken <- round(end.time.full - start.time.full, 2)
full_time.taken


message("##########################################
You have completed Cohorting!!! Woohooo!!! \n You also have saved your data \n Please proceed on to 02_statistical_analysis script \n 
FYI you can JUST run the 02_stats script if no changes are made to the cohort script in the future!!
##########################################")
```
